[Elastic Agent] Fix invalid log level sent to endpoint (#25854)
[Elastic Agent] Fix invalid log level sent to endpoint (#25854)
