+++ x-pack/elastic-agent/pkg/agent/application/gateway/fleet/fleet_gateway_test.go
@@ -123,7 +123,8 @@ func withGateway(agentInfo agentInfo, settings *fleetGatewaySettings, fn withGat
 
 		ctx, cancel := context.WithCancel(context.Background())
 
-		stateStore, err := store.NewStateStore(log, storage.NewDiskStore(paths.AgentStateStoreFile()))
+		diskStore := storage.NewDiskStore(paths.AgentStateStoreFile())
+		stateStore, err := store.NewStateStore(log, diskStore)
 		require.NoError(t, err)
 
 		gateway, err := newFleetGatewayWithScheduler(
@@ -256,7 +257,9 @@ func TestFleetGateway(t *testing.T) {
 
 		ctx, cancel := context.WithCancel(context.Background())
 		log, _ := logger.New("tst", false)
-		stateStore, err := store.NewStateStore(log, storage.NewDiskStore(paths.AgentStateStoreFile()))
+
+		diskStore := storage.NewDiskStore(paths.AgentStateStoreFile())
+		stateStore, err := store.NewStateStore(log, diskStore)
 		require.NoError(t, err)
 
 		gateway, err := newFleetGatewayWithScheduler(
@@ -346,7 +349,8 @@ func TestFleetGateway(t *testing.T) {
 		ctx, cancel := context.WithCancel(context.Background())
 		log, _ := logger.New("tst", false)
 
-		stateStore, err := store.NewStateStore(log, storage.NewDiskStore(paths.AgentStateStoreFile()))
+		diskStore := storage.NewDiskStore(paths.AgentStateStoreFile())
+		stateStore, err := store.NewStateStore(log, diskStore)
 		require.NoError(t, err)
 
 		gateway, err := newFleetGatewayWithScheduler(
