+++ x-pack/elastic-agent/pkg/core/logger/logger.go
@@ -23,6 +23,9 @@ import (
 
 const agentName = "elastic-agent"
 
+// DefaultLogLevel used in agent and its processes.
+const DefaultLogLevel = logp.InfoLevel
+
 // Logger alias ecslog.Logger with Logger.
 type Logger = logp.Logger
 
@@ -93,7 +96,7 @@ func toCommonConfig(cfg *Config) (*common.Config, error) {
 func DefaultLoggingConfig() *Config {
 	cfg := logp.DefaultConfig(logp.DefaultEnvironment)
 	cfg.Beat = agentName
-	cfg.Level = logp.InfoLevel
+	cfg.Level = DefaultLogLevel
 	cfg.ToFiles = true
 	cfg.Files.Path = paths.Logs()
 	cfg.Files.Name = agentName
